+++ aeronautics-addon/common/src/main/java/dev/dsddr/aeronauticsaddon/AeronauticsAddonClient.java
@@ -1,0 +1,10 @@
+package dev.dsddr.aeronauticsaddon;
+
+import dev.dsddr.aeronauticsaddon.content.ponder.AddonPonderPlugin;
+import net.createmod.ponder.foundation.PonderIndex;
+
+public class AeronauticsAddonClient {
+    public static void init() {
+        PonderIndex.addPlugin(new AddonPonderPlugin());
+    }
+}